requirements.txt
streamlit>=1.32.0
Pillow